% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qc.r
\name{calcSD}
\alias{calcSD}
\title{Summarize standard deviation}
\usage{
calcSD(dataset)
}
\arguments{
\item{dataset}{The cleaned input data to be clustered in either data.frame or matrix format.  This should contain only numeric data and must not have any factors, strings, NAs, etc.  It also should not contain any irrelevant columns such as observation ID or redundant data.}
}
\value{
The function returns a double containing the approximated standard deviation of the entire data set.
}
\description{
Summarize the standard deviation of a data set with multiple columns into a single value.  This function accomplishes this task by finding the length of a vector containing the standard deviation of each column.  This function may be useful in establishing a \code{sigma} for qc().
}
\examples{
data(iris)
iris_data <- iris[, -5] # Drop the species factor
sigma <- 0.4 * calcSD(iris_data)
}

